refactor l8l1c
