+++ src/layer/Landsat8AWSLayer.ts
@@ -1,47 +1,6 @@
 import { DATASET_AWS_L8L1C } from './dataset';
-import { AbstractSentinelHubV3WithCCLayer } from './AbstractSentinelHubV3WithCCLayer';
-import { Link, LinkType } from './const';
+import { AbstractLandsat8Layer } from './AbstractLandsat8Layer';
 
-export class Landsat8AWSLayer extends AbstractSentinelHubV3WithCCLayer {
+export class Landsat8AWSLayer extends AbstractLandsat8Layer {
   public readonly dataset = DATASET_AWS_L8L1C;
-
-  protected getTileLinks(tile: Record<string, any>): Link[] {
-    return [
-      {
-        target: tile.dataUri,
-        type: LinkType.AWS,
-      },
-      {
-        target: `${tile.dataUri}_thumb_small.jpg`,
-        type: LinkType.PREVIEW,
-      },
-    ];
-  }
-
-  protected extractFindTilesMeta(tile: any): Record<string, any> {
-    return {
-      ...super.extractFindTilesMeta(tile),
-      sunElevation: tile.sunElevation,
-    };
-  }
-
-  protected extractFindTilesMetaFromCatalog(feature: Record<string, any>): Record<string, any> {
-    return {
-      ...super.extractFindTilesMetaFromCatalog(feature),
-      sunElevation: feature.properties['view:sun_elevation'],
-    };
-  }
-
-  protected getTileLinksFromCatalog(feature: Record<string, any>): Link[] {
-    const { assets } = feature;
-    let result: Link[] = super.getTileLinksFromCatalog(feature);
-
-    if (assets.data && assets.data.href) {
-      result.push({
-        target: assets.data.href.replace('/index.html', `/${feature.id}_thumb_small.jpg`),
-        type: LinkType.PREVIEW,
-      });
-    }
-    return result;
-  }
 }
